part4/requirements.txt
Jinja2
PyYAML
napalm